kelime/requirements.txt
Python==3.6.5
Django==2.2.2
djangorestframework==3.9.4